+++ src/main/scala/edu/ie3/simona/config/SimonaConfig.scala
@@ -119,7 +119,7 @@ object SimonaConfig {
       input: InputConfig,
       output: OutputConfig,
       powerflow: Simona.Powerflow,
-      runtime: RuntimeConfig,
+      runtime: RuntimeConfig = RuntimeConfig(),
       simulationName: String,
       time: Simona.Time = Simona.Time(),
   ) derives ConfigConvert